Reorganize ffi code
